Tercera practica integradora

+++ src/dao/dbController/controller.products.js
@@ -7,6 +7,8 @@ const productSearch = require('../products.dao')
 const adminAccess = require('../../middlewares/adminAcces.middleware')
 const userAcces = require('../../middlewares/userAcces.middleware')
 const ProductsRepository = require('../repository/products.repository')
+const logger = require('../../config/logs/logger.config')
+const ErrorRepository = require('../repository/errors.repository')
 
 
 // Utiliza el middleware de acceso privado para verificar que el usuaio este autenticado si no lo redirecciona al login
@@ -22,9 +24,14 @@ router.get('/', privateAccess, async (req, res, next) => {
     // parsear el objeto con el id del usuario a cadena
     const cartId = cart._id.toString()
     const products = await productSearch(req, message, cartId)
+
+
+    logger.info('Productos cargados con exito', products)
+
     //renderizamos la vista handlebars y pasamos los datos con los que trabajaremos
     res.render('products.handlebars', products);
   } catch (error) {
+    logger.error('Error al cargar los productos', error)
     next(error)
   }
 });
@@ -37,6 +44,7 @@ router.get('/mockingProducts',userAcces, async (req, res, next) => {
     const mockProducts = await productsRepository.generateMockProducts()
     res.json({Productos: mockProducts})
   } catch (error) {
+    logger.error('Error al generar los productos', error)
     next(error)
   }
 })
@@ -45,29 +53,64 @@ router.get('/mockingProducts',userAcces, async (req, res, next) => {
 //Agregar un nuevo producto a la db
 router.post('/', adminAccess , async (req, res, next) => {
   try {
+
+
+    if(req.session.user.role !== 'premium' && req.session.user.role !== 'administrador'){
+      throw new ErrorRepository('Rol de usuario rechazado', 401)
+    }
+
+    if(req.body.owner === null){
+      req.body.owner === 'administrador'
+    }
+
+    req.body.owner = req.session.user.email
+
     const newProduct = await Products.create(req.body)
+    logger.info('Se agrego un producto a la db', newProduct)
     res.json({message: newProduct})
   } catch (error) {
+    logger.error('Error al agregar producto', error)
     next(error)
   }
 })
 
+//Actualizar un producto en especifico
 router.put('/:productId', adminAccess , async (req, res, next) => {
   try {
+
+
+    const product = await Products.findById(req.params.productId)
+    const user = req.session.user
+
+    if (user.role === 'administrador' || (user.email !== 'premium' && product.owner !== 'premium')) {
+      return new ErrorRepository('No tienes permiso para modificar este producto', 401)
+    }
+
     const updatedProduct = await Products.findByIdAndUpdate(req.params.productId, req.body, { new: true })
+    logger.info('Producto actualizado con exito', updatedProduct)
     res.json({ message: 'Product updated successfully', product: updatedProduct })
   } catch (error) {
-    console.log(error)
+    logger.error('Error al actualizar el producto')
     next(error)
   }
 })
 
 router.delete('/:productId', adminAccess , async (req, res, next) => {
   try {
+
+    const product = await Products.findById(req.params.productId)
+    const user = req.session.user
+
+    if (user.role === 'administrador' || (user.email !== 'premium' && product.owner !== 'premium')) {
+      return new ErrorRepository('No tienes permiso para eliminar este producto', 401)
+    }
+
+
     await Products.findByIdAndDelete(req.params.productId)
+    logger.info('Producto eliminado', req.params.productId)
     res.json({message: `Product with ID ${req.params.productId} has been deleted`})
   } catch (error) {
-    console.log(error)
+    logger.error('Error al eliminar el producto', error)
     next(error)
   }
 })
